fix: Refactored Tfl enums into their own namespace
