--- app/graphql/types/category_type.rb
@@ -1,11 +1,0 @@
-module Types
-  class CategoryType < Types::BaseEnum
-    value 'UNDEFINED'
-    value 'REALTIME'
-    value 'PLANNEDWORK'
-    value 'INFORMATION'
-    value 'EVENT'
-    value 'CROWDING'
-    value 'STATUSALERT'
-  end
-end
