--- app/graphql/types/date_time_type.rb
@@ -1,6 +1,0 @@
-module Types
-  class DateTimeType < Types::BaseEnum
-    value 'ARRIVING'
-    value 'DEPARTING'
-  end
-end
