--- app/graphql/types/route_type.rb
@@ -1,12 +1,0 @@
-module Types
-  class RouteType < Types::BaseEnum
-    value 'ALL'
-    value 'UNKNOWN'
-    value 'CYCLE_SUPERHIGHWAYS'
-    value 'QUIETWAYS'
-    value 'CYCLEWAYS'
-    value 'MINI_HOLLANDS'
-    value 'CENTRAL_LONDON_GRID'
-    value 'STREETSPACE_ROUTE'
-  end
-end
