--- app/graphql/types/sky_direction_description_type.rb
@@ -1,12 +1,0 @@
-module Types
-  class SkyDirectionDescriptionType < Types::BaseEnum
-    value 'NORTH'
-    value 'NORTHEAST'
-    value 'EAST'
-    value 'SOUTHEAST'
-    value 'SOUTH'
-    value 'SOUTHWEST'
-    value 'WEST'
-    value 'NORTHWEST'
-  end
-end
