--- app/graphql/types/status_type.rb
@@ -1,11 +1,0 @@
-module Types
-  class StatusType < Types::BaseEnum
-    value 'UNKNOWN'
-    value 'ALL'
-    value 'OPEN'
-    value 'IN_PROGRESS'
-    value 'PLANNED'
-    value 'PLANNED___SUBJECT_TO_FEASIBILITY_AND_CONSULTATION_'
-    value 'NOT_OPEN'
-  end
-end
