+++ app/graphql/types/tfl/entities/identifier_type.rb
@@ -6,8 +6,8 @@ module Types
         field :fullName, String, null: true
         field :id, String, null: true
         field :name, String, null: true
-        field :routeType, Types::RouteType, null: true
-        field :status, Types::StatusType, null: true
+        field :routeType, Types::Tfl::Enums::RouteType, null: true
+        field :status, Types::Tfl::Enums::StatusType, null: true
         field :type, String, null: true
         field :uri, String, null: true
       end
