+++ app/graphql/types/tfl/entities/instruction_step_type.rb
@@ -11,9 +11,9 @@ module Types
         field :longitude, Float, null: true
         field :pathAttribute, Types::Tfl::Entities::PathAttributeType, null: true
         field :skyDirection, Integer, null: true
-        field :skyDirectionDescription, Types::SkyDirectionDescriptionType, null: true
+        field :skyDirectionDescription, Types::Tfl::Enums::SkyDirectionDescriptionType, null: true
         field :streetName, String, null: true
-        field :trackType, Types::TrackType, null: true
+        field :trackType, Types::Tfl::Enums::TrackType, null: true
         field :turnDirection, String, null: true
       end
     end
