+++ app/graphql/types/tfl/entities/journey_planner/search_criteria_type.rb
@@ -4,7 +4,7 @@ module Types
       module JourneyPlanner
         class SearchCriteriaType < Types::BaseObject
           field :dateTime, String, null: true
-          field :dateTimeType, Types::DateTimeType, null: true
+          field :dateTimeType, Types::Tfl::Enums::DateTimeType, null: true
           field :timeAdjustments, Types::Tfl::Entities::JourneyPlanner::TimeAdjustmentsType, null: true
         end
       end
