+++ app/graphql/types/tfl/enums/category_type.rb
@@ -1,0 +1,15 @@
+module Types
+  module Tfl
+    module Enums
+      class CategoryType < Types::BaseEnum
+        value 'UNDEFINED'
+        value 'REALTIME'
+        value 'PLANNEDWORK'
+        value 'INFORMATION'
+        value 'EVENT'
+        value 'CROWDING'
+        value 'STATUSALERT'
+      end
+    end
+  end
+end
