+++ app/graphql/types/tfl/enums/date_time_type.rb
@@ -1,0 +1,10 @@
+module Types
+  module Tfl
+    module Enums
+      class DateTimeType < Types::BaseEnum
+        value 'ARRIVING'
+        value 'DEPARTING'
+      end
+    end
+  end
+end
