+++ app/graphql/types/tfl/enums/route_type.rb
@@ -1,0 +1,16 @@
+module Types
+  module Tfl
+    module Enums
+      class RouteType < Types::BaseEnum
+        value 'ALL'
+        value 'UNKNOWN'
+        value 'CYCLE_SUPERHIGHWAYS'
+        value 'QUIETWAYS'
+        value 'CYCLEWAYS'
+        value 'MINI_HOLLANDS'
+        value 'CENTRAL_LONDON_GRID'
+        value 'STREETSPACE_ROUTE'
+      end
+    end
+  end
+end
