+++ app/graphql/types/tfl/enums/sky_direction_description_type.rb
@@ -1,0 +1,16 @@
+module Types
+  module Tfl
+    module Enums
+      class SkyDirectionDescriptionType < Types::BaseEnum
+        value 'NORTH'
+        value 'NORTHEAST'
+        value 'EAST'
+        value 'SOUTHEAST'
+        value 'SOUTH'
+        value 'SOUTHWEST'
+        value 'WEST'
+        value 'NORTHWEST'
+      end
+    end
+  end
+end
