--- app/graphql/types/track_type.rb
@@ -1,12 +1,0 @@
-module Types
-  class TrackType < Types::BaseEnum
-    value 'CYCLESUPERHIGHWAY'
-    value 'CANALTOWPATH'
-    value 'QUIETROAD'
-    value 'PROVISIONFORCYCLISTS'
-    value 'BUSYROADS'
-    value 'NONE'
-    value 'PUSHBIKE'
-    value 'QUIETWAY'
-  end
-end
