+++ spec/graphql/types/tfl/enums/datetime_type_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe Types::DateTimeType do
+describe Types::Tfl::Enums::DateTimeType do
   subject { described_class }
 
   it { is_expected.to respond_to(:values) }
